# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface VisionStatus {
  /** True when at least one provider API key is configured */
  hasKeys: boolean;
}

export interface ChecklistItem {
  id: string;
  name: string;
}

export interface ReferencePhoto {
  /** Item id this reference photo belongs to */
  id: string;
  /** Base64-encoded JPEG (no data URI prefix) */
  data: string;
}

export interface VisionInspectRequest {
  /**
     * Base64-encoded JPEG belongings photos (no data URI prefix)
     * @minItems 1
     */
  photos: string[];
  /** @minItems 1 */
  items: ChecklistItem[];
  refs?: ReferencePhoto[];
  /** Item ids previously confirmed as false-missing (judge leniently) */
  lenient?: string[];
  /**
     * Optional user-supplied Gemini API key. When either user key is present, only user keys are used (no server-key fallback).
     * @maxLength 300
     */
  geminiKey?: string;
  /**
     * Optional user-supplied Anthropic API key
     * @maxLength 300
     */
  anthropicKey?: string;
}

export interface VisionInspectResult {
  /** Ids of checklist items not visible in any photo */
  missing: string[];
}

export interface SuggestExtra {
  name: string;
  emoji: string;
}

export interface VisionSuggestRequest {
  title: string;
  time?: string;
  location?: string;
  /** @minItems 1 */
  candidates: ChecklistItem[];
  /**
     * Optional user-supplied Gemini API key. When either user key is present, only user keys are used (no server-key fallback).
     * @maxLength 300
     */
  geminiKey?: string;
  /**
     * Optional user-supplied Anthropic API key
     * @maxLength 300
     */
  anthropicKey?: string;
}

export interface VisionSuggestResult {
  ids: string[];
  extra: SuggestExtra[];
}

export interface VisionError {
  error: string;
}