# lib/api-zod/src/generated/types/visionInspectRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChecklistItem } from './checklistItem';
import type { ReferencePhoto } from './referencePhoto';

export interface VisionInspectRequest {
  /**
     * Base64-encoded JPEG belongings photos (no data URI prefix)
     * @minItems 1
     */
  photos: string[];
  /** @minItems 1 */
  items: ChecklistItem[];
  refs?: ReferencePhoto[];
  /** Item ids previously confirmed as false-missing (judge leniently) */
  lenient?: string[];
  /**
     * Optional user-supplied Gemini API key. When either user key is present, only user keys are used (no server-key fallback).
     * @maxLength 300
     */
  geminiKey?: string;
  /**
     * Optional user-supplied Anthropic API key
     * @maxLength 300
     */
  anthropicKey?: string;
}