# lib/api-zod/src/generated/types/visionSuggestRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChecklistItem } from './checklistItem';

export interface VisionSuggestRequest {
  title: string;
  time?: string;
  location?: string;
  /** @minItems 1 */
  candidates: ChecklistItem[];
  /**
     * Optional user-supplied Gemini API key. When either user key is present, only user keys are used (no server-key fallback).
     * @maxLength 300
     */
  geminiKey?: string;
  /**
     * Optional user-supplied Anthropic API key
     * @maxLength 300
     */
  anthropicKey?: string;
}